refactor(onboard): extract web search provider policy
