+++ crates/daemon/src/lib.rs
@@ -90,6 +90,7 @@ mod onboard_finalize;
 mod onboard_preflight;
 pub mod onboard_presentation;
 mod onboarding_model_policy;
+mod onboard_web_search;
 mod provider_credential_policy;
 mod provider_model_probe_policy;
 pub mod provider_presentation;
